requirements.txt
discord.py==1.5.1
python-dotenv==0.4.0
python3-bs4==4.9.2